refactor: make TraceMockTrait
